pass typechecker, remove TrainingManager, fix RL loop
