+++ requirements.txt
@@ -1,2 +1,3 @@
 google-cloud-storage==2.7.0
 bitstring
+sgfmill
